requirements.txt
tensorflow-gpu
numpy
tqdm
matplotlib
Pillow
nltk
scikit-learn

